+++ src/posh-dynargs.psm1
@@ -4,19 +4,30 @@ if (Get-Module posh-dynargs) { return }
 . $PSScriptRoot\Utils.ps1
 
 function Register-LocalArgumentCompleters() {
+    if ($PoshDynargsSettings.EnableTiming) {
+        $sw = [System.Diagnostics.Stopwatch]::StartNew()
+    }
     $argumentCompleterRegistryPath = ".\.argument-completer-registry.json"
     if (!(Test-Path $argumentCompleterRegistryPath)) { return }
-    if ($DynamicArgumentCompleterSettings.LastRegisteredDirectory -eq $PWD.Path) { return }
+    if ($PoshDynargsSettings.LastRegisteredDirectory -eq $PWD.Path) { return }
 
     $completableCommands = ((Get-Content $argumentCompleterRegistryPath | ConvertFrom-Json).completableCommands).name;
     if ($null -eq $completableCommands) { return }
 
     & $PSScriptRoot\Register-DynamicArgumentCompleters.ps1 -commandsToComplete $completableCommands
-    $DynamicArgumentCompleterSettings.LastRegisteredDirectory = $PWD.Path
+    $PoshDynargsSettings.LastRegisteredDirectory = $PWD.Path
+
+    # If timing enabled, display elapsed milliseconds
+    if ($PoshDynargsSettings.EnableTiming) {
+        $sw.Stop()
+        $elapsed = $sw.ElapsedMilliseconds
+        Write-Host "[${elapsed}ms]" -NoNewline -ForegroundColor DarkGray
+    }
 }
 
-$DynamicArgumentCompleterSettings = @{
+$PoshDynargsSettings = @{
     LastRegisteredDirectory = $null;
+    EnableTiming = $false;
 }
 
 if ($GitPromptSettings) {
@@ -28,19 +39,19 @@ else {
     # TODO: Don't really want to clobber the existing prompt.
     #       Consider wrapping the existing prompt.
     # TODO: Bypass overriding prompt via environment setting.
+    $promptScriptBlock = {
+        $origLastExitCode = $global:LASTEXITCODE
 
-    $promptScriptBlock = [scriptblock]::Create(@'
-    $origLastExitCode = $global:LASTEXITCODE
+        Register-LocalArgumentCompleters
 
-    Register-LocalArgumentCompleters
+        $pathInfo = $ExecutionContext.SessionState.Path.CurrentLocation
+        $currentPath = if ($pathInfo.Drive) { $pathInfo.Path } else { $pathInfo.ProviderPath }
 
-    $pathInfo = $ExecutionContext.SessionState.Path.CurrentLocation
-    $currentPath = if ($pathInfo.Drive) { $pathInfo.Path } else { $pathInfo.ProviderPath }
+        $global:LASTEXITCODE = $origLastExitCode
 
-    $global:LASTEXITCODE = $origLastExitCode
+        "$currentPath> "
+    }
 
-    "$currentPath> "
-'@)
     Set-Item Function:\prompt -Value $promptScriptBlock
 }
 
@@ -50,7 +61,7 @@ $exportModuleMemberParams = @{
         'Add-PoshDynargsToProfile'
     );
     Variable = @(
-        'DynamicArgumentCompleterSettings'
+        'PoshDynargsSettings'
     );
 }