+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # posh-dynargs Release History
 
-## 0.1.0 - January 31, 2019
+## 0.1.0.1 - January 12, 2019
 
-- Initial Release
+- Begin documentation, with a note about perf.
+- Rename `$DynamicArgumentCompleterSettings` to `$PoshDynArgsSettings`
+- Add ability to time `Register-LocalArgumentCompleters` (`$PoshDynArgsSettings.EnableTiming = $true`)
+
+## 0.1.0.0 - January 11, 2019
+
+- Initial Release. Just a rough sketch.